+++ src/app/item-triage/ItemTriage.tsx
@@ -1,18 +1,37 @@
+// THIS FILE IS A MESS in order to effectively exclude its content from the bundle
+
 import { settingsSelector } from 'app/dim-api/selectors';
 import BungieImage from 'app/dim-ui/BungieImage';
 import CollapsibleTitle from 'app/dim-ui/CollapsibleTitle';
 import { StatHashListsKeyedByDestinyClass, StatTotalToggle } from 'app/dim-ui/CustomStatTotal';
+import ElementIcon from 'app/dim-ui/ElementIcon';
 import { ExpandableTextBlock } from 'app/dim-ui/ExpandableTextBlock';
-import { ArmorSlotSpecificModSocketIcon } from 'app/dim-ui/SpecialtyModSlotIcon';
+import PressTip from 'app/dim-ui/PressTip';
+import {
+  ArmorSlotSpecificModSocketIcon,
+  SpecialtyModSlotIcon,
+} from 'app/dim-ui/SpecialtyModSlotIcon';
+import { getArmorSvgIcon, getWeaponSvgIcon } from 'app/dim-ui/svgs/itemCategory';
 import { t } from 'app/i18next-t';
 import { allItemsSelector } from 'app/inventory/selectors';
+import PlugTooltip from 'app/item-popup/PlugTooltip';
 import { editLoadout } from 'app/loadout-drawer/LoadoutDrawer';
 import { loadoutsSelector } from 'app/loadout-drawer/selectors';
 import { ItemFilter } from 'app/search/filter-types';
 import { filterFactorySelector } from 'app/search/search-filter';
+import { nameFilter, quoteFilterString } from 'app/search/search-filters/freeform';
+import {
+  classFilter,
+  damageFilter,
+  itemCategoryFilter,
+  itemTypeFilter,
+} from 'app/search/search-filters/known-values';
+import { modslotFilter } from 'app/search/search-filters/sockets';
 import { setSearchQuery } from 'app/shell/actions';
 import { AppIcon, editIcon, searchIcon, thumbsUpIcon } from 'app/shell/icons';
 import { RootState } from 'app/store/types';
+import { getInterestingSocketMetadatas } from 'app/utils/item-utils';
+import { getWeaponArchetype, getWeaponArchetypeSocket } from 'app/utils/socket-utils';
 import { wishListSelector } from 'app/wishlists/selectors';
 import { DestinyClass } from 'bungie-api-ts/destiny2';
 import clsx from 'clsx';
@@ -21,15 +40,19 @@ import { useDispatch, useSelector } from 'react-redux';
 import { DimItem } from '../inventory/item-types';
 // eslint-disable-next-line css-modules/no-unused-class
 import popupStyles from '../item-popup/ItemDescription.m.scss';
-import {
-  factorComboCategories,
-  FactorComboCategory,
-  factorCombos,
-  getItemFactorComboDisplays,
-} from './factors';
 // eslint-disable-next-line css-modules/no-unused-class
 import styles from './ItemTriage.m.scss';
-import { getValueColors } from './ValueDial';
+
+/** whether an item's popup should contain the triage tab */
+export function doShowTriage(item: DimItem) {
+  return (
+    item.destinyVersion === 2 &&
+    (item.bucket.inArmor ||
+      (item.bucket.sort === 'Weapons' &&
+        item.bucket.type !== 'SeasonalArtifacts' &&
+        item.bucket.type !== 'Class'))
+  );
+}
 
 export function ItemTriage({ item }: { item: DimItem }) {
   const dispatch = useDispatch();
@@ -64,51 +87,52 @@ export function ItemTriage({ item }: { item: DimItem }) {
 
   return (
     <div className={styles.itemTriagePane}>
-      <CollapsibleTitle
-        title={'Wishlists'}
-        sectionId={'triage-wishlist'}
-        defaultCollapsed={false}
-        extra={wishlistItem && <AppIcon className="thumbs-up" icon={thumbsUpIcon} />}
-      >
-        {wishlistItem?.notes?.length && (
-          <ExpandableTextBlock linesWhenClosed={3} className={popupStyles.description}>
-            <span className={popupStyles.wishListLabel}>
-              {t('WishListRoll.WishListNotes', { notes: '' })}
-            </span>
-            <span className={popupStyles.wishListTextContent}>{wishlistItem.notes}</span>
-          </ExpandableTextBlock>
-        )}
-      </CollapsibleTitle>
-
+      {item.bucket.inWeapons && (
+        <CollapsibleTitle
+          title={'Wishlists'}
+          sectionId={'triage-wishlist'}
+          defaultCollapsed={false}
+          extra={wishlistItem ? <AppIcon className="thumbs-up" icon={thumbsUpIcon} /> : '–'}
+          disabled={!wishlistItem}
+        >
+          {wishlistItem?.notes?.length && (
+            <ExpandableTextBlock linesWhenClosed={3} className={popupStyles.description}>
+              <span className={popupStyles.wishListLabel}>
+                {t('WishListRoll.WishListNotes', { notes: '' })}
+              </span>
+              <span className={popupStyles.wishListTextContent}>{wishlistItem.notes}</span>
+            </ExpandableTextBlock>
+          )}
+        </CollapsibleTitle>
+      )}
       <CollapsibleTitle
         title={'In Loadouts'}
         sectionId={'triage-loadout'}
         defaultCollapsed={true}
         extra={inLoadouts.length}
         extraOnlyCollapsed
+        disabled={!inLoadouts.length}
       >
-        {inLoadouts.length > 0 && (
-          <ul>
-            {inLoadouts.map((l) => (
-              <li className={styles.loadoutRow} key={l.id}>
-                {l.name}{' '}
-                <a
-                  className={styles.lowKeyButton}
-                  title={t('Loadouts.Edit')}
-                  onClick={() => editLoadout(l, { isNew: false })}
-                >
-                  <AppIcon icon={editIcon} />
-                </a>
-              </li>
-            ))}
-          </ul>
-        )}
+        <ul>
+          {inLoadouts.map((l) => (
+            <li className={styles.loadoutRow} key={l.id}>
+              {l.name}{' '}
+              <a
+                className={styles.lowKeyButton}
+                title={t('Loadouts.Edit')}
+                onClick={() => editLoadout(l, { isNew: false })}
+              >
+                <AppIcon icon={editIcon} />
+              </a>
+            </li>
+          ))}
+        </ul>
       </CollapsibleTitle>
       <CollapsibleTitle
         title={'Item Count'}
         sectionId={'triage-itemcount'}
         defaultCollapsed={false}
-        extra={itemFactors && Math.min(...itemFactors.map((f) => f.count))}
+        extra={itemFactors?.length ? Math.min(...itemFactors.map((f) => f.count)) : undefined}
         extraOnlyCollapsed
       >
         <div className={styles.ownershipTable}>
@@ -207,7 +231,152 @@ export function ItemTriage({ item }: { item: DimItem }) {
   );
 }
 
-function getSimilarItems(
+/** a factor of interest */
+interface Factor {
+  id: string;
+  /** bother checking this factor, if the seed item returns truthy */
+  runIf(item: DimItem): unknown;
+  render(item: DimItem): React.ReactElement | null;
+  filter(item: DimItem): string;
+}
+// factors someone might value in an item, like its mod slot or its element
+const itemFactors: Record<string, Factor> = {
+  class: {
+    id: 'class',
+    runIf: () => true,
+    render: () => null,
+    // render: (item) => (
+    //   <PressTip elementType="span" tooltip={item.classTypeNameLocalized}>
+    //     <ClassIcon classType={item.classType} className={styles.classIcon} />
+    //   </PressTip>
+    // ),
+    filter: classFilter.fromItem!,
+  },
+  name: {
+    id: 'name',
+    runIf: (item) => item.bucket.inWeapons,
+    render: (item) => (
+      <>
+        <BungieImage className={styles.inlineIcon} src={item.icon} /> {item.name}
+      </>
+    ),
+    filter: nameFilter.fromItem!,
+  },
+  element: {
+    id: 'element',
+    runIf: (item) => item.element,
+    render: (item) => (
+      <PressTip elementType="span" tooltip={item.element?.displayProperties.name}>
+        <ElementIcon className={clsx(styles.inlineIcon, styles.smaller)} element={item.element} />
+      </PressTip>
+    ),
+    filter: damageFilter.fromItem!,
+  },
+  weaponType: {
+    id: 'weaponType',
+    runIf: (item) => item.bucket.inWeapons,
+    render: (item) => {
+      const weaponIcon = getWeaponSvgIcon(item);
+      return weaponIcon ? (
+        <PressTip elementType="span" tooltip={item.typeName}>
+          <img
+            className={clsx(styles.inlineIcon, styles.smaller, styles.weaponSvg)}
+            src={getWeaponSvgIcon(item)}
+          />
+        </PressTip>
+      ) : (
+        <>{item.typeName}</>
+      );
+    },
+    filter: itemCategoryFilter.fromItem!,
+  },
+  specialtySocket: {
+    id: 'specialtySocket',
+    runIf: getInterestingSocketMetadatas,
+    render: (item) => (
+      <SpecialtyModSlotIcon className={styles.inlineIcon} item={item} lowRes onlyInteresting />
+    ),
+    filter: modslotFilter.fromItem!,
+  },
+  armorSlot: {
+    id: 'armorSlot',
+    runIf: (item) => item.bucket.inArmor,
+    render: (item) => (
+      <PressTip elementType="span" tooltip={item.typeName}>
+        <img src={getArmorSvgIcon(item)} className={clsx(styles.inlineIcon, styles.weaponSvg)} />
+      </PressTip>
+    ),
+    filter: itemTypeFilter.fromItem!,
+  },
+  archetype: {
+    id: 'archetype',
+    runIf: (item) => getWeaponArchetype(item),
+    render: (item) => {
+      const archetypeSocket = getWeaponArchetypeSocket(item);
+      return (
+        <>
+          {archetypeSocket?.plugged && (
+            <PressTip
+              elementType="span"
+              tooltip={<PlugTooltip item={item} plug={archetypeSocket.plugged} />}
+            >
+              <BungieImage
+                className={styles.inlineIcon}
+                src={archetypeSocket.plugged.plugDef.displayProperties.icon}
+              />
+            </PressTip>
+          )}
+        </>
+      );
+    },
+    filter: (item) =>
+      `perkname:${quoteFilterString(getWeaponArchetype(item)!.displayProperties.name)}`,
+  },
+};
+
+// which factors to check for which buckets
+const factorCombos = {
+  Weapons: [
+    [itemFactors.element, itemFactors.weaponType],
+    [itemFactors.archetype, itemFactors.weaponType],
+  ],
+  Armor: [
+    [itemFactors.class, itemFactors.element, itemFactors.specialtySocket, itemFactors.armorSlot],
+    [itemFactors.class, itemFactors.element, itemFactors.specialtySocket],
+    [itemFactors.name],
+  ],
+  General: [[itemFactors.element]],
+};
+type FactorComboCategory = keyof typeof factorCombos;
+const factorComboCategories = Object.keys(factorCombos);
+
+function getItemFactorComboDisplays(exampleItem: DimItem) {
+  if (!exampleItem.bucket.sort || !factorComboCategories.includes(exampleItem.bucket.sort)) {
+    return [];
+  }
+  return factorCombos[exampleItem.bucket.sort as FactorComboCategory]
+    .filter((factorCombo) => factorCombo.every((factor) => factor.runIf(exampleItem)))
+    .map((factorCombo) => renderFactorCombo(exampleItem, factorCombo));
+}
+
+function renderFactorCombo(exampleItem: DimItem, factorCombo: Factor[]) {
+  return (
+    <div className={styles.factorCombo}>
+      {factorCombo.map((factor) => (
+        <React.Fragment key={factor.id}>{factor.render(exampleItem)}</React.Fragment>
+      ))}
+    </div>
+  );
+}
+
+/** returns [dark, light] variations along a 1->100  red->yellow->green line */
+function getValueColors(value: number): [string, string] {
+  const hue = value * 1.25;
+  const light = Math.floor(-(value ** 2 / 250) + (2 * value) / 5 + 30);
+  return [`hsl(${hue}, 100%, 8%)`, `hsl(${hue}, 100%, ${light}%)`];
+}
+
+export function getSimilarItems(
   exampleItem: DimItem,
   allItems: DimItem[],
   filterFactory: (query: string) => ItemFilter
@@ -274,7 +443,7 @@ const notabilityThreshold = 0.8;
 /**
  * returns an entry for each notable stat found on the seed item
  */
-function getNotableStats(
+export function getNotableStats(
   exampleItem: DimItem,
   customTotalStatsByClass: StatHashListsKeyedByDestinyClass,
   allItems: DimItem[]
