--- src/app/item-triage/ValueDial.m.scss.d.ts
@@ -1,9 +1,0 @@
-// This file is automatically generated.
-// Please do not change this file!
-interface CssExports {
-  'junkIcon': string;
-  'keepIcon': string;
-  'tagIcon': string;
-}
-export const cssExports: CssExports;
-export default cssExports;
